Improve denest and fix testing infrastructure.
Denest transform now works for a single nested structure. It should
hopefully place the denesting links in their LCA position, which is on
the top-most graph node created by denesting, but that needs to be
tested.

Testing infrastructure computes structure ids correctly and more simply
so numbers aren't skipped. I was double-counting the map size.