Fix off-by-one error in path computation for LCA.
LCA wasn't using the root node in the path and wasn't adding the
bottom-most node either. This meant that if a link was already in LCA
position with source and target immediately below LCA, the algorithm
would think the two nodes were on the same ancestor, not its immediate
children. Self links are currently not supported by LCA so this resulted
in an error.