Switch flow to separate tags from UIDs.
This decoupling allows us to more closely model the program config
hierarchy in Sidewinder, allowing for more continuous visualizations.
For instance, operators in FFS contain their arguments as children, even
though operators do not really own their arguments and the animation
of operators and their arguments should be separated. See
InheritExamples.re for a worked example of how to use the new system to
circumvent this issue. NOTE: This version doesn't support external
functions properly, since external functions lose the guarantee that all
nodes involved in a given flow have the same structure, but we assume
the structures are the same in this implementation to properly propagate
flow from Dummy nodes to Leaf nodes.